requirements.txt
xrpl-py==2.0.0
python-dotenv==1.0.0
pandas==2.1.4
